feat: Refactor layout components to import ChatBot from a specific path for consistency
